DAS-2267: Switch trajectory-subsetter over to use the shared_utilies
